requirements/co2mpas-freeze-AIO-UP-1.7.3.post2.dev0.txt
alabaster==0.7.10
asteval==0.9.8
Babel==2.4.0
beautifulsoup4==4.5.3
boltons==17.1.0
Bottleneck==1.2.1
bs4==0.0.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
co2mpas==1.7.3
colorama==0.3.9
cycler==0.10.0
decorator==4.0.11
dill==0.2.6
docopt==0.6.2
docutils==0.13.1
easygui==0.98.1
et-xmlfile==1.0.1
Flask==0.12.2
formulas==0.0.9
future==0.16.0
gitdb2==2.0.0
GitPython==2.1.5
graphviz==0.8.3
idna==2.5
imagesize==0.7.1
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
jsonschema==2.6.0
llvmlite==0.23.0
lmfit==0.9.5
logutils==0.3.4.1
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.0.0
networkx==2.1
numba==0.38.0
numexpr==2.6.5
numpy==1.11.3+mkl
openpyxl==2.5.3
packaging==17.1
pandalone==0.2.4
pandas==0.23.0
parsedatetime==2.4
Pillow==4.0.0
pycparser==2.17
Pygments==2.2.0
pykalman==0.9.5
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.6.0
python-gnupg==0.4.1
pytz==2016.10
PyYAML==3.12
rainbow-logging-handler==2.2.2
regex==2018.02.21
requests==2.18.1
schedula==0.1.18
schema==0.6.6
scikit-learn==0.19.1
scipy==0.19.0
six==1.10.0
smmap2==2.0.1
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinx-rtd-theme==0.2.4
toolz==0.8.2
tornado==4.4.2
tqdm==4.11.2
transitions==0.5.2
Unidecode==0.4.21
urllib3==1.21.1
validate-email==1.3
Werkzeug==0.12.2
wltp==0.1.0a2
xgboost==0.71
xlrd==1.1.0
XlsxWriter==0.9.5
